Separated out code into separate files, used common substream_summary.
